requirements.txt
altgraph==0.17.4
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
bidict==0.22.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.2.0
cx-Oracle==8.3.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
Flask==3.0.0
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
keyboard==0.13.5
lxml==4.9.3
MarkupSafe==2.1.3
mouse==0.7.1
MouseInfo==0.1.3
numpy==1.26.2
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.3
pefile==2023.2.7
Pillow==9.5.0
psutil==5.9.6
psycopg2==2.9.9
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyinstaller-versionfile==2.1.1
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytweening==1.0.7
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pywinauto==0.6.8
PyYAML==6.0.1
requests==2.31.0
selenium==4.15.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
trio==0.23.1
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
zope.event==5.0
zope.interface==6.1
